+++ needle/core/framework/cli.py
@@ -42,6 +42,7 @@ class CLI(Framework):
         self._init_home()
         self.show_banner()
         self.do_reload(None)
+        self._history_load()
 
     # ==================================================================================================================
     # INIT METHODS
@@ -58,7 +59,7 @@ class CLI(Framework):
         self.register_option('app', '', False, 'Bundle ID of the target application (e.g., com.example.app). Leave empty to launch wizard')
         self.register_option('setup_device', Constants.GLOBAL_SETUP_DEVICE, True, 'Set to true to enable auto-configuration of the device (installation of all the tools needed)')
         self.register_option('output_folder', Constants.GLOBAL_OUTPUT_FOLDER, True, 'Full path of the output folder, where to store the output of the modules')
-        #self.register_option('proxy', None, False, 'Proxy server (address:port)')
+        self.register_option('save_history', Constants.GLOBAL_SAVE_HISTORY, True, 'Persists command history across sessions')
 
     def _init_global_vars(self):
         # Setup Printer